third_party/blink/web_tests/fast/events/keydown-function-keys-expected.txt

This test verifies the keycodes sent when we press function keys.

Type=keydown,keyCode=112,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=112,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=113,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=113,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=114,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=114,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=115,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=115,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=116,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=116,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=117,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=117,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=118,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=118,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=119,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=119,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=120,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=120,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=121,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=121,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=122,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=122,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keydown,keyCode=123,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
Type=keyup,keyCode=123,ctrlKey=false,shiftKey=false,altKey=false,metaKey=false
